Prepare v3.0 release of Joveler.Compression family

+++ Joveler.Compression.XZ.Tests/XZStreamTests.cs
@@ -83,7 +83,7 @@ namespace Joveler.Compression.XZ.Tests
                 };
                 XZThreadedCompressOptions threadOpts = new XZThreadedCompressOptions
                 {
-                    Threads = (uint)threads,
+                    Threads = threads,
                 };
 
                 string sampleFile = Path.Combine(TestSetup.SampleDir, sampleFileName);
